+++ packages/commerce-sdk-react/src/hooks/ShopperLogin/queryKeyHelpers.ts
@@ -4,10 +4,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import type {ShopperLogin} from 'commerce-sdk-isomorphic'
+import {ShopperLogin} from 'commerce-sdk-isomorphic'
 import {Argument, ExcludeTail} from '../types'
-import {getCustomKeys, pick} from '../utils'
-import paramKeysMap from './paramKeys'
+import {pickValidParams} from '../utils'
 
 // We must use a client with no parameters in order to have required/optional match the API spec
 type Client = ShopperLogin<{shortCode: string}>
@@ -53,8 +52,10 @@ export const getUserInfo: QueryKeyHelper<'getUserInfo'> = {
         '/oauth2/userinfo'
     ],
     queryKey: (params: Params<'getUserInfo'>) => {
-        const paramKeys = [...paramKeysMap['getUserInfo'], ...getCustomKeys(params)]
-        return [...getUserInfo.path(params), pick(params, paramKeys)]
+        return [
+            ...getUserInfo.path(params),
+            pickValidParams(params, ShopperLogin.paramKeys.getUserInfo)
+        ]
     }
 }
 
@@ -66,12 +67,10 @@ export const getWellknownOpenidConfiguration: QueryKeyHelper<'getWellknownOpenid
         '/oauth2/.well-known/openid-configuration'
     ],
     queryKey: (params: Params<'getWellknownOpenidConfiguration'>) => {
-        const paramKeys = [
-            ...paramKeysMap['getWellknownOpenidConfiguration'],
-            ...getCustomKeys(params)
+        return [
+            ...getWellknownOpenidConfiguration.path(params),
+            pickValidParams(params, ShopperLogin.paramKeys.getWellknownOpenidConfiguration)
         ]
-
-        return [...getWellknownOpenidConfiguration.path(params), pick(params, paramKeys)]
     }
 }
 
@@ -83,7 +82,9 @@ export const getJwksUri: QueryKeyHelper<'getJwksUri'> = {
         '/oauth2/jwks'
     ],
     queryKey: (params: Params<'getJwksUri'>) => {
-        const paramKeys = [...paramKeysMap['getJwksUri'], ...getCustomKeys(params)]
-        return [...getJwksUri.path(params), pick(params, paramKeys)]
+        return [
+            ...getJwksUri.path(params),
+            pickValidParams(params, ShopperLogin.paramKeys.getJwksUri)
+        ]
     }
 }