+++ packages/commerce-sdk-react/src/hooks/ShopperStores/index.test.ts
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2023, Salesforce, Inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import {ShopperStores} from 'commerce-sdk-isomorphic'
+import {getUnimplementedEndpoints} from '../../test-utils'
+import * as queries from './query'
+
+describe('Shopper Search hooks', () => {
+    test('all endpoints have hooks', () => {
+        const unimplemented = getUnimplementedEndpoints(ShopperStores, queries)
+        expect(unimplemented).toEqual([])
+    })
+})